fix: dialog disappear when resizing but focus was trapped
This is more of a workaround. Ideally we would redraw the dialog using
the new dimensions. We need to start passing the Resize event downstream
for this to happen.

The workaround just closes any open dialogs during a Resize event, this
way the focus goes back to the editor. Odd, but doesn't result in an
invalid editor state.